reorganize folder - auth/auth-controller.js   fix config reference - config/config.js -> config.js   moved filed - config/express_ionic.js   delete - config/passport.js   delete - middleware/auth.js   new auth middleware - config/express.js -> middleware/express.js   moved to middleware folder - user/user-model.js   fix config reference
