+++ server/middleware/auth.js
@@ -1,0 +1,33 @@
+const userModel = require('../user/user-model');
+
+const isAuthenticated = (req, res, next) => {
+  if (!(req.headers && req.headers.authorization)) {
+    return res.status(400).json({
+      status: 'Please log in'
+    });
+  }
+  // decode the token
+  var header = req.headers.authorization.split(' ');
+  var token = header[1];
+  const payload = jwt.decode(token, config.token.secret);
+  localAuth.decodeToken(token, (err, payload) => {
+    if (err) {
+      return res.status(401).json({
+        status: 'Token has expired'
+      });
+    } else {
+      // check if the user still exists in the db
+      return knex('users').where({id: parseInt(payload.sub)}).first()
+      .then((user) => {
+        next();
+      })
+      .catch((err) => {
+        res.status(500).json({
+          status: 'error'
+        });
+      });
+    }
+  });
+}
+
+module.exports = isAuthenticated;
